+++ src/features/reminders/index.ts
@@ -1,0 +1,2 @@
+export * from "./localReminderResolver";
+export * from "./components/LocalReminderPanel";